add controller Test, moved all students' activity to the Test controller
