+++ app/helpers/tasks_helper.rb
@@ -1,19 +1,6 @@
 #coding: utf-8
 module TasksHelper
 
-  def bnf_elements_lists
-    return {["<падеж>", "padezh"] => Padezh_list,
-            ["<род>", "gender"] => Gender_list,
-            ["<число>", "number"] => Number_list,
-            ["<одушевленность>", "odush"] => Odush_list,
-            ["<вид>", "vid"] => Vid_list,
-            ["<время>", "tense"] => Tense_list,
-            ["<имя семантической валентности>", "sem_val"] => Sem_val_list,
-            ["<предлог>", "pred"] => Pred_list,
-            ["<семантический признак>", "sem_prizn"] => Sem_prizn_list,
-            ["<кодификатор части речи>", "codificator"] => Codificator_list}
-  end
-
   def display_standard_table(columns, collection = {})
 
     thead = content_tag :thead do
