+++ app/helpers/test_helper.rb
@@ -1,0 +1,2 @@
+module TestHelper
+end
