--- app/views/tasks/get_result.html.erb
@@ -1,9 +1,0 @@
-<% content_for(:head) do %>
-
-  <title>Результаты работы</title>
-
-<% end %>
-
-<div id="result">
-  <span>Временная заглушка</span>
-</div>
