+++ app/views/tasks/index.html.erb
@@ -4,7 +4,7 @@
 
 <% end %>
 
-<%= link_to "Создать новое задание", tasks_new_path %>
+<%= link_to "Создать новое задание", new_task_path%>
 <%= display_standard_table([{name: "id", display_name: "Номер"},
 							{name: "sentence1", display_name: "Задание"},
 							{name: "reference", display_name: ""}], @tasks) %>
