+++ app/views/test/get_g.html.erb
@@ -21,7 +21,7 @@
 	</div>
 </div>
 <div id="answer_send">
-	<form id="answer" method="POST" action="next_component" >
+	<form id="answer" method="POST" action= "<%= next_component_path %>" >
 		<%= tag(:input, :type => "hidden", :id => "answer_content", :name => "g_answer") %>
 		<%= tag(:input, :type => "hidden", :name => request_forgery_protection_token.to_s, :value => form_authenticity_token) %>
 		<%= tag(:input, :type => "hidden", :name => "component", :value => "G") %>
