+++ app/views/test/get_s.html.erb
