+++ app/views/test/get_v.html.erb
@@ -14,7 +14,7 @@
 </div>
 <div id="answer_send">
   <%= form_tag(next_component_path, method: "POST", id:"answer") do %>
-		<%= hidden_field_tag(:answer_content, name: "v_answer_bnf") %>
+		<%= hidden_field_tag(:answer_content) %>
 		<%= tag(:input, type: "hidden", name: "component", value: "V") %>
 		<input type="submit" onmousedown="sendAnswer();" />
   <% end %>
