+++ db/seeds.rb
@@ -18,3 +18,5 @@ k6_224 = Group.create(number: 'K6-224')
 st = k6_224.students.create(fio: "Test Student")
 testuser = st.create_user(login: 'student0', pass: Digest::MD5.hexdigest('lolka'),
                       role: 'student')
+admin = User.create(login: 'admin', pass: Digest::MD5.hexdigest('admin'),
+                   role: 'admin')
