+++ test/controllers/test_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class TestControllerTest < ActionController::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
