+++ test/helpers/test_helper_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class TestHelperTest < ActionView::TestCase
+end
